feat: update product pricing to use PriceInOere for consistency across models and frontend

+++ Webshop.Data/Models/CartItem.cs
@@ -6,6 +6,6 @@ namespace Webshop.Data.Models
         public int ProductId { get; set; }
         public required string ProductName { get; set; }
         public int Quantity { get; set; }
-        public long PriceInOere { get; set; }
+        public int PriceInOere { get; set; }
     }
 }
